FIX: Provide data storage functions
